servercontainer/requirements.txt
amqp==2.6.1
asgiref==3.3.4
async-timeout==4.0.2
attrs==21.2.0
billiard==3.6.4.0
certifi==2021.10.8
chardet==4.0.0
click==8.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
DateTimeRange==1.2.0
defusedxml==0.7.1
Deprecated==1.2.13
diff-match-patch==20200713
Django==3.2.4
django-3-jet==1.0.7
django-cors-headers==3.7.0
django-debug-toolbar==3.2.4
django-elasticsearch-dsl==7.2.2
django-elasticsearch-dsl-drf==0.22.4
django-filter==2.4.0
django-geojson==3.2.0
django-import-export==2.5.0
django-ipware==4.0.0
django-jet==1.0.8
django-leaflet==0.28.1
django-nine==0.2.5
django-redis==5.2.0
django-seed==0.3.1
djangorestframework==3.12.4
djangorestframework-gis==0.17
djangorestframework-recursive==0.1.2
djangorestframework-simplejwt==5.0.0
drf-spectacular==0.17.2
elasticsearch==7.17.0
elasticsearch-dsl==7.4.0
et-xmlfile==1.1.0
Faker==13.2.0
feedparser==6.0.6
future==0.18.2
google-api-python-client==1.4.1
httplib2==0.19.1
inflection==0.5.1
jsonschema==3.2.0
kombu==4.6.11
llvmlite==0.39.0
Markdown==3.3.4
MarkupPy==1.14
mbstrdecoder==1.1.0
oauth2client==4.1.3
odfpy==1.4.1
openpyxl==3.0.7
packaging==21.3
pandas==1.4.3
Pillow==9.0.1
prompt-toolkit==3.0.29
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==2.3.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2022.1
PyYAML==5.4.1
pyzk==0.9
redis==4.2.0
reportlab==3.6.8
rsa==4.7.2
schedule==1.1.0
sgmllib3k==1.0.0
six==1.16.0
sqlparse==0.4.1
tablib==3.0.0
toposort==1.7
typepy==1.3.0
typing_extensions==4.1.1
uritemplate==3.0.1
urllib3==1.26.8
vine==1.3.0
wcwidth==0.2.5
wrapt==1.14.0
xlrd==2.0.1
xlwt==1.3.0
gunicorn
celery
mysqlclient
